feat(ui): add pages module and move pages into submodules
