+++ ui/src/pages/mod.rs
@@ -1,0 +1,5 @@
+mod hello;
+mod profile;
+
+pub use hello::HelloPage;
+pub use profile::ProfilePage;
